feat: implement full backend installers
