+++ installers/__init__.py
@@ -1,0 +1,6 @@
+"""Backend-specific installer modules for the Tomex project."""
+
+# Re-export install functions for convenience
+from . import windows, wsl, docker, pip_installer
+
+__all__ = ["windows", "wsl", "docker", "pip_installer"]
